New Feature: user defined colors for target and saving interactive heatmap
New argument for the plots: cmap_target_dict. It expects a dictionary
with the exhaustive target: color mapping. Color could be str or tuple, which is
understandable by matplotlib.colors. It plots the colors on
feat_importance bar plots (top left corner) and also on the static and
interactive classification heatmap. Also, if save is given for
interactive heatmap, it saves to a html file instead of throwing an
exception. Not implemented for regression type heatmap. If
cmap_target_dict is None, it plots the default colors ("Blues_r" for count
plot and "Greens" for heatmap). Tested, works.